Move Rust integrations into core crate
