+++ litellm-rust/crates/ai-gateway/src/constants.rs
@@ -5,26 +5,6 @@
 //! modules. Env-overridable tunables keep their `DEFAULT_*` value here; the env
 //! read + fallback happens at the host/config layer.
 
-/// Default LiteLLM control-plane base URL for request-log egress when
-/// `LITELLM_PROXY_BASE_URL` is unset.
-pub(crate) const DEFAULT_PROXY_BASE_URL: &str = "http://localhost:4000";
-
-/// The logs ingest path appended to the proxy base. Not a tunable; it is the
-/// proxy's API contract (the rust-control-plane router on the Python proxy).
-pub(crate) const RUST_CONTROL_PLANE_LOGS_PATH: &str = "/v1/rust_control_plane/logs";
-
-/// Default bounded channel depth for the log-egress worker.
-/// Override: `LITELLM_LOG_CHANNEL_CAPACITY`.
-pub(crate) const DEFAULT_CHANNEL_CAPACITY: usize = 4096;
-
-/// Default max records POSTed per request to the control plane.
-/// Override: `LITELLM_LOG_BATCH_SIZE`.
-pub(crate) const DEFAULT_MAX_BATCH_SIZE: usize = 256;
-
-/// Default partial-batch flush cadence, in ms.
-/// Override: `LITELLM_LOG_FLUSH_INTERVAL_MS`.
-pub(crate) const DEFAULT_FLUSH_INTERVAL_MS: u64 = 500;
-
 /// Provider attributed to realtime sessions in the logging payload.
 #[cfg(feature = "server")]
 pub(crate) const DEFAULT_PROVIDER: &str = "openai";
