+++ litellm-rust/crates/core/Cargo.toml
@@ -7,9 +7,11 @@ repository.workspace = true
 
 [dependencies]
 rand.workspace = true
+reqwest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
+tokio = { workspace = true, features = ["sync", "time"] }
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
