+++ litellm-rust/crates/core/src/integrations/types.rs
